testing rr3db configuration

+++ src/optical_flow/models/raft/utils.py
@@ -8,7 +8,7 @@ def bilinear_sampler(img, coords, mask=False):
     xgrid = 2*xgrid/(W-1) - 1
     ygrid = 2*ygrid/(H-1) - 1
 
-    grid = torch.cat([xgrid, ygrid], dim=-1)
+    grid = torch.cat([xgrid, ygrid], dim=-1).type_as(img)
     img = F.grid_sample(img, grid, align_corners=True)
 
     if mask:
